+++ src/frontend/src/app/list-tasks/list-tasks.component.css
@@ -1,0 +1,7 @@
+.card-header {
+    background-color: whitesmoke;
+}
+
+table.scroll {
+    height: 300px;
+}
