+++ src/frontend/src/app/scheduled-process/scheduled-process.component.html
@@ -51,41 +51,6 @@
 
   <div class="tile">
     <div class="tile is-parent">
-      <div class="tile is-child card">
-        <header class="card-header">
-          <div class="card-header-title">
-            Tasks &nbsp;
-            <div class="select" (change)="getTasks($event.target.value)">
-              <select ng-model="selectedTaskType" id="task-state">
-                <option>Latest</option>
-                <option>Running</option>
-              </select>
-            </div>
-          </div>
-        </header>
-        <div class="card-content">
-          <table class="table is-striped  scroll is-responsive">
-            <thead>
-              <tr>
-                <th>Process</th>
-                <th>State</th>
-                <th>Start time</th>
-                <th>End Time</th>
-                <th>Exit Code</th>
-                <th>Reason</th>
-              </tr>
-            </thead>
-            <tr *ngFor="let task of tasksData">
-              <td><small>{{task.process_name}}</small></td>
-              <td><small>{{task.state}}</small></td>
-              <td><small>{{task.start_time != 'None' ?(task.start_time | date: 'dd-MM-yyyy HH:mm:ss') : task.start_time}}</small></td>
-              <td><small>{{task.end_time != 'None'?(task.end_time | date: 'dd-MM-yyyy HH:mm:ss') : task.end_time}}</small></td>
-              <td><small>{{task.exit_code}}</small></td>
-              <td><small>{{task.reason}}</small></td>
-              <td><button *ngIf="'RUNNING' == task.state" class="button is-small is-danger" (click)="cancelRunninTask(task.id)">Cancel</button></td>
-            </tr>
-          </table>
-        </div>
-      </div>
+     <app-list-tasks></app-list-tasks>
     </div>
   </div>
